Move Manifests interface to manifests package
This commit moves the interface `Manifests` so it is defined alongside
the procedures which depend upon it.

Secondarily, it moves the AppendManifestToBuffer procedure into the
Manifests interface, so to avoid the assumption that manifests are
always YAML.

To break a circular dependency, it was necessary to put cluster.Mock
in its own package -- probably for the best anyway, since it ought to
(continue to) rely only on exported types.